refactor: split worker and core entrypoints
