+++ wrangler.toml
@@ -1,5 +1,5 @@
 name = "echo-pdf"
-main = "src/index.ts"
+main = "src/worker.ts"
 compatibility_date = "2026-03-06"
 
 [assets]
